Move sidebar to PaneHolder
