+++ Files/IShellPage.cs
@@ -25,6 +25,7 @@ namespace Files
 
         public abstract void Refresh_Click();
         public void UpdatePathUIToWorkingDirectory(string newWorkingDir, string singleItemOverride = null);
+        public void NavigateToPath(string navigationPath, Type sourcePageType);
     }
 
     public interface IPaneHolder : IDisposable
